Split ping types one-per-file + DocC (cmux Aziz policy)
- CmxRoutePing.swift -> CmxRoutePingResult.swift / CmxRoutePinging.swift /
  CmxNetworkRoutePinger.swift (one major type each).
- Add DocC to CmxNetworkRoutePinger.init/ping.
- Move PingTestListener to its own test file; drop the tautological
  fake-pinger test (it tested Swift, not our code), fold isReachable/isListening
  assertions into the refused test instead.
